working version

+++ examples/miplib/targets.py
@@ -10,4 +10,4 @@ class TargetsTrait:
     def get_targets_from_batch_as_one_hot(self, batch):
         counts = self.get_targets_from_batch_as_count(batch)
         one_hot = F.one_hot(counts, num_classes=self.num_classes)
-        return one_hot # shape (batch_size, num_classes)
+        return one_hot.float() # shape (batch_size, num_classes)
